feature 2548 multisize - custom sizes restricted to those requested by theme/plugin code refacto
git-svn-id: http://piwigo.org/svn/trunk@13021 68402e56-0260-453c-a942-63ccdbb3a9ee

+++ admin/include/functions.php
@@ -202,15 +202,18 @@ SELECT
     }
 
     $ok = true;
-    foreach ($files as $path)
-    {
-      if (is_file($path) and !unlink($path))
-      {
-        $ok = false;
-        trigger_error('"'.$path.'" cannot be removed', E_USER_WARNING);
-        break;
-      }
-    }
+		if (!isset($conf['never_delete_originals']))
+		{
+			foreach ($files as $path)
+			{
+				if (is_file($path) and !unlink($path))
+				{
+					$ok = false;
+					trigger_error('"'.$path.'" cannot be removed', E_USER_WARNING);
+					break;
+				}
+			}
+		}
 
     if ($ok)
     {
